package moved, consumer interface
