+++ src/main/java/org/eocqrs/kafka/Consumer.java
@@ -1,0 +1,9 @@
+package org.eocqrs.kafka;
+
+public interface Consumer<X> {
+
+  void consume(int partition, Data<X> message);
+
+  Data<X> dataize();
+
+}
