+++ src/main/java/org/eocqrs/kafka/data/KfDataized.java
@@ -20,39 +20,22 @@
  * SOFTWARE.
  */
 
-package org.eocqrs.kafka;
+package org.eocqrs.kafka.data;
 
-import com.jcabi.xml.XML;
-import java.util.HashMap;
-import java.util.Map;
 import lombok.RequiredArgsConstructor;
-import org.apache.kafka.clients.producer.KafkaProducer;
-import org.eocqrs.kafka.xml.TextXpath;
+import org.eocqrs.kafka.Dataized;
 
 /**
  * @author Aliaksei Bialiauski (abialiauski.dev@gmail.com)
  * @since 0.0.0
  */
 @RequiredArgsConstructor
-public final class KfProducerSettings<K, X> implements ProducerSettings<K, X> {
+public final class KfDataized<X> implements Dataized<X> {
 
-  private final XML xml;
+  private final X data;
 
   @Override
-  public KafkaProducer<K, X> producer() {
-    final Map<String, Object> config = new HashMap<>(3);
-    config.put("bootstrap.servers",
-      new TextXpath(this.xml, "//bootstrapServers")
-        .toString()
-    );
-    config.put("key.serializer",
-      new TextXpath(this.xml, "//keySerializer")
-        .toString()
-    );
-    config.put("value.serializer",
-      new TextXpath(this.xml, "//valueSerializer")
-        .toString()
-    );
-    return new KafkaProducer<>(config);
+  public X dataize() {
+    return this.data;
   }
 }
