+++ src/main/java/org/eocqrs/kafka/package-info.java
@@ -26,5 +26,4 @@
  * @author Aliaksei Bialiauski (abialiauski.dev@gmail.com)
  * @since 0.0.0
  */
-
 package org.eocqrs.kafka;
